+++ WolvenKit.Cache/CollisionCache/CollisionCacheItem.cs
@@ -93,8 +93,11 @@ namespace WolvenKit.Cache
     {
         public IWitcherArchive Bundle { get; set; }
         public string Name { get; set; }
-        public long Size { get; set; }
+        public uint Size { get; set; }
         public uint ZSize { get; set; }
+        /// <summary>
+        /// !!! Double check when writing !!! Some files use 64bit, older files may use 32bit.
+        /// </summary>
         public long PageOffset { get; set; }
         public string CompressionType => "Zlib";
         public uint NameOffset;
